Dampen taquilla swings and isolate VIP from walk-up substitution
Walk-up attendance now references a baseline-pricing holder count and
moves only by 40% of the holder delta, so small season-ticket price
changes no longer whipsaw projected matchday revenue. Premium zones
(VIP/palco) are excluded from the substitution model entirely — those
seats sell as season-long contracts and don't have a walk-up market,
so their holder swings now leave taquilla flat.

Move the season-ticket-editor Alpine component into a dedicated JS
module and add a loading indicator to the Ingresos del Estadio card so
users see when the live preview is recalculating in the background.
